alternate awk actions based on after flag
Simply expanding the $after variable even when it was set to an empty string was causing awk to insert an extra space after commented out lines, which in turn caused the tests to fail because the character sequences weren't identical. The fix is to supply alternate versions of the comment handling action, one for cases where the --after flag is being used and another for cases where it is omitted.