concise logging
Longer error messages are pleasant to read and more informative, but restricting the script's printed output to only include the compiled filenames makes it possible to effectively run a Markdown file directly by using a subshell. This is too powerful a feature to ignore! It may be a good idea to add more verbose logging back in using an argument.