change example program from JavaScript to Ruby
The previous version of this example was written in literate-style JavaScript, which of course is an extremely popular language. However, I made the mistake of simply alert()-ing a few strings, which only works in the browser, and not in the command line contexts in which this script is most likely to be used at first, immediately after download/installation. Instead, the example program should be something that quickly creates command-line output so new users can see how it works. This could be accomplished by just switching the alert() function to console.log(), but even that would require the user to install Node.js before it would execute. Python would be a great choice, since it is installed by default on most OS X and Linux systems, but the print command changed from a statement into a function between versions 2 and 3, and which version you're going to get on any given system is still a bit of a crap shoot, so it's not possible to predict whether the user will need it to use parentheses. That leaves us with Ruby, which is similarly widespread but doesn't have the syntax schism around printing to the terminal. Granted, puts is perhaps not the ideal description of that functionality to present to a novice (print is clearer, for example), but it the language has a very clean syntax and is fairly popular language in CS education, so here we are.