explain puts method in example program
Ruby's puts method is a little esoteric and might be confusing to newcomers, or at least more so than Python's print or JavaScript's console.log -- but we can explain it! In fact, that is the entire point of this tool! So here's a short explanation to make the example program a little easier to understand for those who don't have any Ruby knowledge.